requirements.txt
streamlit
pyinstaller
pyyaml
pandas
numpy
altair
plotly
